move LegalizeJSInterace to src/passes
