+++ src/asm2wasm.h
@@ -84,7 +84,6 @@ Name I64("i64"),
      I64_BC2I("i64_bc2i"),
      I64_CTTZ("i64_cttz"),
      I64_CTLZ("i64_ctlz"),
-     TEMP_RET_0("tempRet0"),
      I64S_REM("i64s-rem"),
      I64U_REM("i64u-rem"),
      I64S_DIV("i64s-div"),
@@ -106,224 +105,6 @@ static void abort_on(std::string why, IString element) {
 Index indexOr(Index x, Index y) {
   return x ? x : y;
 }
-
-Expression* recreateI64(Builder& builder, Expression* low, Expression* high) {
-  return
-    builder.makeBinary(
-      OrInt64,
-      builder.makeUnary(
-        ExtendUInt32,
-        low
-      ),
-      builder.makeBinary(
-        ShlInt64,
-        builder.makeUnary(
-          ExtendUInt32,
-          high
-        ),
-        builder.makeConst(Literal(int64_t(32)))
-      )
-    )
-  ;
-};
-
-Expression* recreateI64(Builder& builder, Index low, Index high) {
-  return recreateI64(builder, builder.makeGetLocal(low, i32), builder.makeGetLocal(high, i32));
-};
-
-Expression* getI64High(Builder& builder, Index index) {
-  return
-    builder.makeUnary(
-      WrapInt64,
-      builder.makeBinary(
-        ShrUInt64,
-        builder.makeGetLocal(index, i64),
-        builder.makeConst(Literal(int64_t(32)))
-      )
-    )
-  ;
-}
-
-Expression* getI64Low(Builder& builder, Index index) {
-  return
-    builder.makeUnary(
-      WrapInt64,
-      builder.makeGetLocal(index, i64)
-    )
-  ;
-}
-
-// i64 imports and exports must be turned into i32, i32
-struct LegalizeJSInterface : public Pass {
-  void run(PassRunner* runner, Module* module) override {
-    // for each illegal export, we must export a legalized stub instead
-    for (auto& ex : module->exports) {
-      if (ex->kind == Export::Function) {
-        // if it's an import, ignore it
-        if (auto* func = module->checkFunction(ex->value)) {
-          if (isIllegal(func)) {
-            auto legalName = makeLegalStub(func, module);
-            ex->value = legalName;
-          }
-        }
-      }
-    }
-    // for each illegal import, we must call a legalized stub instead
-    std::vector<Import*> newImports; // add them at the end, to not invalidate the iter
-    for (auto& im : module->imports) {
-      if (im->kind == Import::Function && isIllegal(im->functionType)) {
-        Name funcName;
-        auto* legal = makeLegalStub(im.get(), module, funcName);
-        illegalToLegal[im->name] = funcName;
-        newImports.push_back(legal);
-      }
-    }
-    if (illegalToLegal.size() > 0) {
-      for (auto* im : newImports) {
-        module->addImport(im);
-      }
-
-      // fix up imports: call_import of an illegal must be turned to a call of a legal
-
-      struct FixImports : public WalkerPass<PostWalker<FixImports, Visitor<FixImports>>> {
-        bool isFunctionParallel() override { return true; }
-
-        Pass* create() override { return new FixImports(illegalToLegal); }
-
-        std::map<Name, Name>* illegalToLegal;
-
-        FixImports(std::map<Name, Name>* illegalToLegal) : illegalToLegal(illegalToLegal) {}
-
-        void visitCallImport(CallImport* curr) {
-          auto iter = illegalToLegal->find(curr->target);
-          if (iter == illegalToLegal->end()) return;
-
-          if (iter->second == getFunction()->name) return; // inside the stub function itself, is the one safe place to do the call
-          replaceCurrent(Builder(*getModule()).makeCall(iter->second, curr->operands, curr->type));
-        }
-      };
-
-      PassRunner passRunner(module);
-      passRunner.add<FixImports>(&illegalToLegal);
-      passRunner.run();
-    }
-  }
-
-private:
-  // map of illegal to legal names for imports
-  std::map<Name, Name> illegalToLegal;
-
-  template<typename T>
-  bool isIllegal(T* t) {
-    for (auto param : t->params) {
-      if (param == i64) return true;
-    }
-    if (t->result == i64) return true;
-    return false;
-  }
-
-  // JS calls the export, so it must call a legal stub that calls the actual wasm function
-  Name makeLegalStub(Function* func, Module* module) {
-    Builder builder(*module);
-    auto* legal = new Function();
-    legal->name = Name(std::string("legalstub$") + func->name.str);
-
-    auto* call = module->allocator.alloc<Call>();
-    call->target = func->name;
-    call->type = func->result;
-
-    for (auto param : func->params) {
-      if (param == i64) {
-        call->operands.push_back(recreateI64(builder, legal->params.size(), legal->params.size() + 1));
-        legal->params.push_back(i32);
-        legal->params.push_back(i32);
-      } else {
-        call->operands.push_back(builder.makeGetLocal(legal->params.size(), param));
-        legal->params.push_back(param);
-      }
-    }
-
-    if (func->result == i64) {
-      legal->result = i32;
-      auto index = builder.addVar(legal, Name(), i64);
-      auto* block = builder.makeBlock();
-      block->list.push_back(builder.makeSetLocal(index, call));
-      if (module->checkGlobal(TEMP_RET_0)) {
-        block->list.push_back(builder.makeSetGlobal(
-          TEMP_RET_0,
-          getI64High(builder, index)
-        ));
-      } else {
-        block->list.push_back(builder.makeUnreachable()); // no way to emit the high bits :(
-      }
-      block->list.push_back(getI64Low(builder, index));
-      block->finalize();
-      legal->body = block;
-    } else {
-      legal->result = func->result;
-      legal->body = call;
-    }
-
-    // a method may be exported multiple times
-    if (!module->checkFunction(legal->name)) {
-      module->addFunction(legal);
-    }
-    return legal->name;
-  }
-
-  // wasm calls the import, so it must call a stub that calls the actual legal JS import
-  Import* makeLegalStub(Import* im, Module* module, Name& funcName) {
-    Builder builder(*module);
-    auto* type = new FunctionType();
-    type->name =  Name(std::string("legaltype$") + im->name.str);
-    auto* legal = new Import();
-    legal->name = Name(std::string("legalimport$") + im->name.str);
-    legal->module = im->module;
-    legal->base = im->base;
-    legal->kind = Import::Function;
-    legal->functionType = type;
-    auto* func = new Function();
-    func->name = Name(std::string("legalfunc$") + im->name.str);
-    funcName = func->name;
-
-    auto* call = module->allocator.alloc<CallImport>();
-    call->target = legal->name;
-
-    for (auto param : im->functionType->params) {
-      if (param == i64) {
-        call->operands.push_back(getI64Low(builder, func->params.size()));
-        call->operands.push_back(getI64High(builder, func->params.size()));
-        type->params.push_back(i32);
-        type->params.push_back(i32);
-      } else {
-        call->operands.push_back(builder.makeGetLocal(func->params.size(), param));
-        type->params.push_back(param);
-      }
-      func->params.push_back(param);
-    }
-
-    if (im->functionType->result == i64) {
-      call->type = i32;
-      Expression* get;
-      if (module->checkGlobal(TEMP_RET_0)) {
-        get = builder.makeGetGlobal(TEMP_RET_0, i32);
-      } else {
-        get = builder.makeUnreachable(); // no way to emit the high bits :(
-      }
-      func->body = recreateI64(builder, call, get);
-      type->result = i32;
-    } else {
-      call->type = im->functionType->result;
-      func->body = call;
-      type->result = im->functionType->result;
-    }
-    func->result = im->functionType->result;
-
-    module->addFunction(func);
-    module->addFunctionType(type);
-    return legal;
-  }
-};
 
 // useful when we need to see our parent, in an asm.js expression stack
 struct AstStackHelper {
@@ -1154,7 +935,7 @@ void Asm2WasmBuilder::processAsm(Ref ast) {
   passRunner.add<AutoDrop>(); // FinalizeCalls may cause us to require additional drops
   if (wasmOnly) {
     // we didn't legalize i64s in fastcomp, and so must legalize the interface to the outside
-    passRunner.add<LegalizeJSInterface>();
+    passRunner.add("legalize-js-interface");
   }
   if (optimize) {
     // autodrop can add some garbage
@@ -1243,8 +1024,8 @@ void Asm2WasmBuilder::processAsm(Ref ast) {
           x64 = Builder::addVar(func, "x64", i64),
           y64 = Builder::addVar(func, "y64", i64);
     auto* body = allocator.alloc<Block>();
-    body->list.push_back(builder.makeSetLocal(x64, recreateI64(builder, xl, xh)));
-    body->list.push_back(builder.makeSetLocal(y64, recreateI64(builder, yl, yh)));
+    body->list.push_back(builder.makeSetLocal(x64, I64Utilities::recreateI64(builder, xl, xh)));
+    body->list.push_back(builder.makeSetLocal(y64, I64Utilities::recreateI64(builder, yl, yh)));
     body->list.push_back(
       builder.makeIf(
         builder.makeGetLocal(r, i32),
@@ -1273,10 +1054,10 @@ void Asm2WasmBuilder::processAsm(Ref ast) {
     body->list.push_back(
       builder.makeSetGlobal(
         tempRet0,
-        getI64High(builder, x64)
+        I64Utilities::getI64High(builder, x64)
       )
     );
-    body->list.push_back(getI64Low(builder, x64));
+    body->list.push_back(I64Utilities::getI64Low(builder, x64));
     body->finalize();
     func->body = body;
   }
