+++ src/passes/passes.h
@@ -27,7 +27,10 @@ Pass *createCoalesceLocalsWithLearningPass();
 Pass *createDeadCodeEliminationPass();
 Pass *createDuplicateFunctionEliminationPass();
 Pass *createExtractFunctionPass();
+Pass *createFullPrinterPass();
+Pass *createLegalizeJSInterfacePass();
 Pass *createLowerIfElsePass();
+Pass *createMinifiedPrinterPass();
 Pass *createMergeBlocksPass();
 Pass *createMetricsPass();
 Pass *createNameListPass();
@@ -35,8 +38,6 @@ Pass *createNameManagerPass();
 Pass *createOptimizeInstructionsPass();
 Pass *createPostEmscriptenPass();
 Pass *createPrinterPass();
-Pass *createMinifiedPrinterPass();
-Pass *createFullPrinterPass();
 Pass *createRelooperJumpThreadingPass();
 Pass *createRemoveImportsPass();
 Pass *createRemoveMemoryPass();
